pyproject.toml
[project]
name = "fast"
version = "0.1.0"
description = ""
authors = [
    {name = "Bno", email = "Breno_live2005@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "redis (>=6.4.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "kafka-python (>=2.3.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=7.1.0,<8.0.0)"
]

[tool.poetry.dev-dependencies]
# Pin pytest to a version compatible with Python 3.9 (pytest 8.x)
pytest = "^8.2.5"

[tool.pytest.init_options]
pythonpath = ["."]